% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{add_unique_ids}
\alias{add_unique_ids}
\title{Add unique ID variables when binding LIS or LWS files.}
\usage{
add_unique_ids(file, file_name, file_level)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{file_level}{A string indicating the level of the file. Valid inputs are:
'household', 'h', 'person' or 'p'.

\lifecycle{experimental}}
}
\description{
Add unique ID variables when binding LIS or LWS files.
}
\keyword{internal}
