% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_from_other_packages.R
\name{all.is.numeric}
\alias{all.is.numeric}
\title{Check if all elements in a vector are numeric}
\usage{
\method{all}{is.numeric}(x, what = c("test", "vector"), extras = c(".", "NA"))
}
\description{
Function from: Frank E Harrell Jr, with contributions from Charles Dupont and many others. (2020). Hmisc: Harrell Miscellaneous. R package version 4.4-1.
https://CRAN.R-project.org/package=Hmisc
}
\keyword{internal}
