% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all.values.equal}
\alias{all.values.equal}
\title{All values are equal
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
\method{all}{values.equal}(target, current)
}
\description{
Like base::all.equal(), but ignoring attributes by default.
}
\keyword{internal}
