% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_lissy_files}
\alias{bind_lissy_files}
\title{Bind multiple LIS or LWS datasets}
\usage{
bind_lissy_files(lissy_files, create_unique_id = TRUE)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{create_unique_id}{Computes new variables 'unique_hid' and 'unique_pid'
as unique row identifiers. Defaults to TRUE.}
}
\value{
A tibble with the stacked LIS or LWS files.
}
\description{
Converts a list of LIS or LWS datasets into a single file. Creates a new 'file' variable with the
name of the dataset and computes unique person and household
identifiers if 'create_unique_id = TRUE' (default).
}
\details{
Adds the attribute 'binded = TRUE' to the resulting tibble.
}
\examples{
\dontrun{
lissy_datasets_as_list <- read_lissy_files(c("fr84h", "fr94h", "fr10h"))
 lissy_datasets_as_tibble <- bind_lissy_files(lissy_datasets = lissy_datasets_as_list, create_unique_id = TRUE)
}
}
