% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_file_name_format}
\alias{change_file_name_format}
\title{Change the format of a file name}
\usage{
change_file_name_format(file_names, to_format)
}
\arguments{
\item{file_names}{A string with the name of the file.}

\item{to_format}{A string with the desired format. This can be one of the following:
\itemize{
\item 'ccyy' Two digit country, two digit year. E.g. "fr84"
\item 'ccyyyy' Two digit country, four digit year. E.g. "fr1984"
\item 'ccyyl' Two digit country, two digit year, one digit level ('h' or 'p' for household/person). E.g. "fr84h"
\item 'ccyyyyl' Two digit country, four digit year, one digit level. E.g. "fr1984h"
\item 'ccyyd' Two digit country, two digit year, one digit database (LIS, LWS or ERFLIS - i/w/e). E.g. "fr84i"
\item 'ccyyyyd' Two digit country, four digit year, one digit database (LIS, LWS or ERFLIS - i/w/e). E.g. "fr1984i"
\item 'ccyydl' Two digit country, two digit year, one digit indicating LIS, LWS or ERFLIS (i/w/e/r), one digit household/person. E.g. "fr84ih"
\item 'ccyyyydl' Two digit country, two digit year,one digit indicating LIS, LWS or ERFLIS (i/w/e/r), one digit household/person. E.g. "fr1984ih"
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Converts a character vector with a file names to a different format.

Is used within 'implement_adjust_by_lisppp()', .
}
\keyword{internal}
