% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dataset_levels.R
\name{check_file_lists_contain_same_datasets}
\alias{check_file_lists_contain_same_datasets}
\title{Checks if two lists have the same datasets}
\usage{
check_file_lists_contain_same_datasets(lissy_hfiles, lissy_pfiles)
}
\description{
To be used only within 'merge_dataset_levels()'
}
\keyword{internal}
