% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_github_version}
\alias{check_github_version}
\title{Retrieve the current version of lissyrtools from the GitHub repository}
\usage{
check_github_version()
}
\value{
A character string indicating the current version on GitHub.
}
\description{
Fetches the latest version of the \code{lissyrtools} package from its DESCRIPTION file on GitHub.
This allows comparison with the locally installed version. Using an outdated version may
result in missing support for the most recent countries, datasets, or variable revisions,
which can cause inaccurate outputs in functions that rely on up-to-date metadata.
}
\examples{
\dontrun{
library(lissyrtools)
library(utils)
library(assertthat)

local_version <- as.character(utils::packageVersion("lissyrtools"))
remote_github_version <- check_github_version()

assertthat::assert_that(local_version == remote_github_version)
}
}
