% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input_in_weight_argument}
\alias{check_input_in_weight_argument}
\title{Validate the Weight Variable Input}
\usage{
check_input_in_weight_argument(wgt_name)
}
\arguments{
\item{weight_var}{A character string of length 1 indicating the name of the weight variable used.}
}
\value{
No return value. The function is used for validation and emits a message if the weight variable name seems unusual.
}
\description{
This function checks the validity of the weight variable provided by the user, issuing a helpful message if the variable name
does not end with "wgt", as expected by LIS conventions.
}
\keyword{internal}
