% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_for_lissyuse.R
\name{check_length_iso2}
\alias{check_length_iso2}
\title{Checks for the length of the characters in the argument \code{data}.}
\usage{
check_length_iso2(data)
}
\arguments{
\item{data}{A character vector.}
}
\value{
Stops if length of some characters is not 2 or 4, and outputs the invalid characters.
}
\description{
Checks the length of the characters in the argument \code{data}.
}
\keyword{internal}
