% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missing_share}
\alias{check_missing_share}
\title{Checks the share of missing values for each column acroos a list of data frames.}
\usage{
check_missing_share(data_list, include_zeros = FALSE)
}
\arguments{
\item{data_list}{A list}

\item{include_zeros}{Logical. If \code{TRUE}, values equal to zero are also accounted in the percentage of missings.}
}
\value{
A list
}
\description{
Checks the share of missing values for each column acroos a list of data frames.
}
\keyword{internal}
