% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{check_variable_level}
\alias{check_variable_level}
\alias{safely_check_variable_level}
\title{Checks if the variable is 'household' or 'person' level.}
\usage{
check_variable_level(variable)

safely_check_variable_level(...)
}
\arguments{
\item{variable}{A String with the name of the variable. E.g. 'pi11', 'dhi'.}
}
\description{
\lifecycle{experimental}
}
\details{
Throws an error if it couldn't be checked.
}
\keyword{internal}
