% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions_helpers.R
\name{checks_compute_functions}
\alias{checks_compute_functions}
\title{Perform checks for 'compute' functions.}
\usage{
checks_compute_functions(file, file_name, variable, weight = NULL)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{variable}{A string with the name of the variable for which checks should be computed.}

\item{weight}{A string with the name of the variable in 'file' that should be
used as sample weights. If NULL (default), checks for weight variable are not performed.}
}
\description{
\lifecycle{experimental}
Perform checks within 'compute_*()' functions. The checks performed are:
\itemize{
\item The file is a data.frame type of object.
\item 'varaible' is a column in file
\item 'variable' is a numeric variable.
\item The weight variable is a column in file.
\item 'weight' variable contains valid values (i.e. different than NAs and 0s).
}
}
\details{
The checks related to the 'weight' variable are only performed if argument
'weight' is not NULL.
}
\keyword{internal}
