% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{compute_atkinson}
\alias{compute_atkinson}
\title{Compute Atkinson index.}
\usage{
compute_atkinson(
  file,
  file_name,
  variable,
  epsilon,
  weight = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{variable}{A string with the name of the variable for which Atkinson should be computed.}

\item{epsilon}{A number with the inequality adversion parameter. Needs to be epsilon > 0.}

\item{weight}{A string with the name of the variable in 'file' that should be
used as sample weights.}

\item{na.rm}{A boolean. Indicates if NAs should be ignored. Defaults to FALSE.}
}
\value{
A numeric vector with the gini index.
}
\description{
\lifecycle{experimental}
Compute the Atkinson index with  weights.
}
