% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{compute_gini}
\alias{compute_gini}
\title{Compute gini index.}
\usage{
compute_gini(file, file_name, variable, weight = NULL, na.rm = FALSE)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{variable}{A string with the name of the variable for which gini should be computed.}

\item{weight}{A string with the name of the variable in 'file' that should be
used as sample weights.}

\item{na.rm}{A boolean. Indicates if NAs should be ignored. Defaults to FALSE.}
}
\value{
A numeric vector with the gini index.
}
\description{
\lifecycle{experimental}
Compute the gini index with  weights.
}
\examples{
\dontrun{
lissy_datasets <- read_lissy_files(c("fr84h", "fr94h", "fr10h"))
compute_gini(lissy_datasets[["fr1984h"]], "dhi", na.rm = TRUE)
}
}
