% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{compute_mean}
\alias{compute_mean}
\alias{compute_median}
\alias{compute_ratio}
\title{Compute standard indicators.}
\usage{
compute_mean(file, file_name, variable, weight = NULL, na.rm = FALSE)

compute_median(file, file_name, variable, weight = NULL, na.rm = FALSE)

compute_ratio(
  file,
  file_name,
  variable,
  ratio = c(0.9, 0.1),
  weight = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{file}{A tibble or data.frame with a LIS or LWS file.}

\item{file_name}{A string with the name of the LIS or LWS file.}

\item{variable}{A string with the name of the variable for which the
indicator should be computed.}

\item{weight}{A string with the name of the variable in 'file' that should be
used as sample weights.}

\item{na.rm}{A boolean. Indicates if NAs should be ignored. Defaults to FALSE.}

\item{ratio}{A vector of two numeric values between 0 and 1. Defines the
percentiles in the numerator and denominator respectively. E.g. (0.9, 0.1)
computes the 90/10 ratio.}
}
\value{
A numeric vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Compute the weighted mean, median or percentile ratios for a variable.
}
\keyword{internal}
