% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{compute_percentiles}
\alias{compute_percentiles}
\title{Compute percentiles.}
\usage{
compute_percentiles(
  file,
  file_name,
  variable,
  breaks = seq(0, 1, 0.1),
  weight = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{variable}{A character vector of length one with the indicator that needs to be transformed.}

\item{breaks}{A numeric vector specifying the percentiles that should be computed. Defaults to all deciles.}

\item{weight}{A string with the name of the variable in 'file' that should be
used as sample weights.}

\item{na.rm}{A boolean indicating if the computation should ignore missing
values in 'variable' and 'weight'.}
}
\value{
A data.frame with two columns: 'percentile' and 'value'. The first contains
the label of the percentiles computed (e.g. '0.5' for median, '0.2' for first quintile).
The second contains the values of these in the distribution of 'variable'.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Compute the  weighted percentages for a LIS or LWS variable.

It uses 'pwgt' or 'hwgt' to weight the indicator.
}
\examples{
\dontrun{
lissy_datasets <- read_lissy_files(c("fr84h", "fr94h", "fr10h"))
compute_percentiles(lissy_datasets[["fr1984h"]], variable = "dhi", na.rm = TRUE)
}
}
\keyword{internal}
