% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{compute_poverty_rate}
\alias{compute_poverty_rate}
\title{Compute poverty rate.}
\usage{
compute_poverty_rate(
  file,
  file_name,
  variable,
  times_median,
  weight = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{variable}{A character vector of length one with the indicator that needs to be transformed.}

\item{times_median}{A number with the factor by which the median should be multiplied
E.g. 0.5 for a poverty rate of 50\% the median.}

\item{weight}{A string with the name of the variable in 'file' that should be
used as sample weights.}

\item{na.rm}{A boolean indicating if the computation should ignore missing
values in 'variable' and 'weight'.}
}
\value{
A tibble with two columns: 'percentile' and 'value'. The first contains
the label of the percentiles computed (e.g. '0.5' for median, '0.2' for first quintile).
The second contains the values of these in the distribution of 'variable'.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Computes the weighted poverty rate for a LIS or LWS variable.

It uses 'pwgt' or 'hwgt' to weight the indicator.
}
\examples{
\dontrun{
lissy_datasets <- read_lissy_files(c("fr84h", "fr94h", "fr10h"))
compute_poverty_rate(lissy_datasets[["fr1984h"]], file_name = "fr84h", variable = "dhi", times_median = 0.5, na.rm = TRUE)
}
}
\keyword{internal}
