% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weighted_functions.R
\name{compute_weighted_percentiles}
\alias{compute_weighted_percentiles}
\title{Title}
\usage{
compute_weighted_percentiles(
  var,
  wgt = NULL,
  probs = seq(0, 1, 0.25),
  na.rm = TRUE,
  share = FALSE
)
}
\arguments{
\item{var}{}

\item{probs}{}

\item{na.rm}{}

\item{share}{}
}
\description{
Title
}
