% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_list_from_ccyy_to_cc_names_yyyy}
\alias{convert_list_from_ccyy_to_cc_names_yyyy}
\title{Rearranges a list whose elements are ccyy's into country names with numeric vectors names after its years.}
\usage{
convert_list_from_ccyy_to_cc_names_yyyy(data_list)
}
\arguments{
\item{data_list}{A list}
}
\value{
A list
}
\description{
Rearranges a list whose elements are ccyy's into country names with numeric vectors names after its years.
}
\keyword{internal}
