% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_inc_ref_year}
\alias{data_inc_ref_year}
\title{LWS income reference years}
\format{
A data frame with 2 variables:
}
\usage{
data_inc_ref_year
}
\description{
A data frame containing the reference years for the income variables for each LWS dataset.
}
\keyword{datasets}
