% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_vars_labels}
\alias{data_vars_labels}
\title{Variable labels}
\format{
A data frame with 2 variables:
}
\usage{
data_vars_labels
}
\description{
A data frame containing all LIS and LWS variables and its labels.
}
\keyword{datasets}
