% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{database_character_to_name}
\alias{database_character_to_name}
\title{Converts 'i', 'w', 'e' into 'LIS', 'LWS' 'ERFLIS'}
\usage{
database_character_to_name(x)
}
\description{
Converts 'i', 'w', 'e' into 'LIS', 'LWS' 'ERFLIS'
}
\keyword{internal}
