% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{define_option}
\alias{define_option}
\title{Define an option}
\usage{
define_option(option_name, value)
}
\arguments{
\item{option_name}{The value of the option.}
}
\description{
This function defines an option by using the name of the new option and its
value. Can be used to define options programmatically.
}
\keyword{internal}
