% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deflators}
\alias{deflators}
\title{CPI and PPP deflators.}
\format{
A data frame with 8 variables:
}
\source{
\url{https://www.lisdatacenter.org/resources/ppp-deflators/}
}
\usage{
deflators
}
\description{
A dataset containing the Consumer Price Index (CPI) and Purchasing Power Parity (PPP)
deflators that can be used to directly compare absolute monetary values across different
LIS, LWS and LCS datasets.
}
\keyword{datasets}
