% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{determine_file_level}
\alias{determine_file_level}
\title{Determine the level of the file}
\usage{
determine_file_level(lissy_files, files_level)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{files_level}{A string indicating the level of the file. Valid inputs are:
'household', 'h', 'person' or 'p'. If NULL (default), the file level will
be retrived from the 'lissy_files' attributes.}
}
\value{
A character vector with the level of the file
}
\description{
Determine the level of the file
}
\keyword{internal}
