% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{determine_weight}
\alias{determine_weight}
\title{Determine the name of the weight variable}
\usage{
determine_weight(lissy_files, variable, files_level, variable_level)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character string indicating the aggregate for which the indicator needs to be computed.}

\item{files_level}{A string indicating the level of the file. Valid inputs are:
'household', 'h', 'person' or 'p'. If NULL (default), the file level will
be retrived from the 'lissy_files' attributes.}

\item{variable_level}{Level of the variable. Should be either 'household', 'h', 'person' or 'p'.
If NULL (default), the function will try to guess the level of the variable.
This is done by comparing the value in 'variable' with pre-set lists of variables.}
}
\value{
A character vector with the name of the weight variable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Computes the name of the weight variable based on the information from the 'lissy_files'
object, 'files_level' and 'variable_level'
}
\keyword{internal}
