% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erflis_household_variables}
\alias{erflis_household_variables}
\title{ERFLIS Household Variables}
\format{
Character vector
}
\usage{
erflis_household_variables
}
\description{
A character vector containing the names of household-level variables used in ERFLIS datasets.
}
\keyword{internal}
