% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{es16_h_lcs}
\alias{es16_h_lcs}
\title{Sample Household-Level LCS Dataset - ES 2016}
\format{
Data frame with 1,000 rows and 552 columns.
}
\usage{
es16_h_lcs
}
\description{
Sample Household-Level LCS Dataset - ES 2016
}
\examples{
\dontrun{
es16_h_lcs
}
}
\keyword{datasets}
