% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{es16_p_lcs}
\alias{es16_p_lcs}
\title{Sample Individual-Level LCS Dataset - ES 2016}
\format{
Data frame with 2,641 rows and 105 columns.
}
\usage{
es16_p_lcs
}
\description{
Sample Individual-Level LCS Dataset - ES 2016
}
\examples{
\dontrun{
es16_p_lcs
}
}
\keyword{datasets}
