% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{es24_h_lcs}
\alias{es24_h_lcs}
\title{Sample Household-Level LCS Dataset - ES 2024}
\format{
Data frame with 1,000 rows and 552 columns.
}
\usage{
es24_h_lcs
}
\description{
Sample Household-Level LCS Dataset - ES 2024
}
\examples{
\dontrun{
es24_h_lcs
}
}
\keyword{datasets}
