% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{es24_p_lcs}
\alias{es24_p_lcs}
\title{Sample Individual-Level LCS Dataset - ES 2024}
\format{
Data frame with 2,533 rows and 105 columns.
}
\usage{
es24_p_lcs
}
\description{
Sample Individual-Level LCS Dataset - ES 2024
}
\examples{
\dontrun{
es24_p_lcs
}
}
\keyword{datasets}
