% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_information_from_file_name}
\alias{extract_information_from_file_name}
\title{Extract information from file name}
\usage{
extract_information_from_file_name(file_name, current_format)
}
\arguments{
\item{file_name}{A string with the name of the file.}

\item{current_format}{A string with the currant format.}
}
\value{
A list with the following elements:
\itemize{
\item 'country_' ('cc')
\item 'year_' ('yyyy')
\item 'database_' ('d')
\item 'level_' ('l')
}
}
\description{
Extracts the information regarding country, year, database and level. The
latter two only if available.
}
\details{
To be used inside 'change_file_name_format()'.
}
\keyword{internal}
