% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{four_digit_year}
\alias{four_digit_year}
\title{Compute four digit year from two digit one.}
\usage{
four_digit_year(two_digit_year)
}
\arguments{
\item{two_digit_year}{An integer or string value with two digits (from 00 to 99).}
}
\value{
A strign with four digit year.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Computes the four digit year from a string with two digits year.
}
\details{
To be used within 'extract_information_from_file_name()'.
}
\keyword{internal}
