% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_and_years.R
\name{get_countries_lcs}
\alias{get_countries_lcs}
\title{Print all the country code in LCS}
\usage{
get_countries_lcs()
}
\value{
A data frame.
}
\description{
Print all the country code in LCS
}
\examples{
get_countries_lcs()
}
