% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_and_years.R
\name{get_countries_lis}
\alias{get_countries_lis}
\title{Print all the country code in LIS}
\usage{
get_countries_lis()
}
\value{
A data frame.
}
\description{
Print all the country code in LIS
}
\examples{
get_countries_lis()
}
