% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_and_years.R
\name{get_countries_lws}
\alias{get_countries_lws}
\title{Print all the country code in LWS}
\usage{
get_countries_lws()
}
\value{
A data frame.
}
\description{
Print all the country code in LWS
}
\examples{
get_countries_lws()
}
