% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_database}
\alias{get_database}
\title{Get the database from a list of files}
\usage{
get_database(lissy_files)
}
\arguments{
\item{lissy_files}{A list of lissy h-level, p-level or merged files.}
}
\description{
Retrieves the 'database' attribute from a list of files.
}
\keyword{internal}
