% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_database_config.R
\name{get_database_config}
\alias{get_database_config}
\title{Get database-specific configuration for lissyuse internals}
\usage{
get_database_config(database)
}
\arguments{
\item{database}{A character value. One of "lis" (default), "lws", or "lcs".}
}
\value{
A list.
}
\description{
Get database-specific configuration for lissyuse internals
}
\keyword{internal}
