% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_functions.R
\name{get_directory}
\alias{get_directory}
\title{Get directory from parameters file}
\usage{
get_directory()
}
\value{
A string with the LIS, LWS or ERFLIS directory.
}
\description{
Get directory from parameters file
}
\keyword{internal}
