% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_index_hh_heads}
\alias{get_index_hh_heads}
\title{Get the index of hh-heads}
\usage{
get_index_hh_heads(file)
}
\arguments{
\item{file}{A tibble or data.frame with a LIS or LWS file.}
}
\value{
A logical vector with TRUE for hh-heads and FALSE otherwise.
}
\description{
Returns a logical vector indicating which rows correspond to household
heads.
}
\keyword{internal}
