% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_lissy_attributes}
\alias{get_lissy_attributes}
\title{Get the lissy attributes}
\usage{
get_lissy_attributes(lissy_object)
}
\arguments{
\item{lissy_object}{A list of tibbles or a tibble.}
}
\value{
A list with three elements.
}
\description{
Retrieves the 'level', 'merged_levels' and 'database' attributes from list of tibbles or a tibble.
}
\keyword{internal}
