% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{get_lws_file_income_reference_year}
\alias{get_lws_file_income_reference_year}
\title{Retrieve the reference year for LWS files}
\usage{
get_lws_file_income_reference_year(file_name)
}
\arguments{
\item{file_name}{String with the name of the dataset in one of the following formats:
\enumerate{
\item 'ccyy' Two digit country, two digit year. E.g. "fr84"
\item 'ccyyyy' Two digit country, four digit year. E.g. "fr1984"
\item 'ccyyl' Two digit country, two digit year, one digit level. E.g. "fr84h"
\item 'ccyyyyl' Two digit country, four digit year, one digit household/person ('h' or 'p'). E.g. "fr1984h"
\item 'ccyydl' Two digit country, two digit year, one digit indicating LIS, LWS or ERFLIS (i/w/e/r), one digit household/person. E.g. "fr84ih"
\item 'ccyyyydl' Two digit country, two digit year,one digit indicating LIS, LWS or ERFLIS (i/w/e/r), one digit household/person. E.g. "fr1984ih"
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
