% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels_notes_surveys.R
\name{get_surveys_lcs}
\alias{get_surveys_lcs}
\title{Print the survey used to construct the LCS datasets for a given country.}
\usage{
get_surveys_lcs(iso2)
}
\arguments{
\item{iso2}{A character vector with valid iso2 codes of countries present in LCS}
}
\value{
A list, made of character vectors. Each elements corresponds to a country in LCS.
}
\description{
Print the survey used to construct the LCS datasets for a given country.
}
\examples{
get_surveys_lcs("it")
get_surveys_lcs(iso2 = c("lu", "es"))
}
