% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels_notes_surveys.R
\name{get_surveys_lis}
\alias{get_surveys_lis}
\title{Print the survey used to construct the LIS datasets for a given country.}
\usage{
get_surveys_lis(iso2)
}
\arguments{
\item{iso2}{A character vector with valid iso2 codes of countries present in LIS}
}
\value{
A list, made of character vectors. Each elements corresponds to a country in LIS.
}
\description{
Print the survey used to construct the LIS datasets for a given country.
}
\examples{
get_surveys_lis("it")
get_surveys_lis(iso2 = c("uy", "pe"))
}
