% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels_notes_surveys.R
\name{get_surveys_lws}
\alias{get_surveys_lws}
\title{Print the survey used to construct the LWS datasets for a given country.}
\usage{
get_surveys_lws(iso2)
}
\arguments{
\item{iso2}{A character vector with valid iso2 codes of countries present in LWS}
}
\value{
A list, made of character vectors. Each elements corresponds to a country in LWS.
}
\description{
Print the survey used to construct the LWS datasets for a given country.
}
\examples{
get_surveys_lws("it")
get_surveys_lws(iso2 = c("fr", "jp"))
}
