% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_and_years.R
\name{get_years_lcs}
\alias{get_years_lcs}
\title{Print all the existing years in LCS for a given country.}
\usage{
get_years_lcs(iso2)
}
\arguments{
\item{iso2}{A character vector with valid iso2 codes of countries present in LCS.}
}
\value{
A list, made of numeric vectors. Each elements corresponds to a country in LCS.
}
\description{
Print all the existing years in LCS for a given country.
}
\examples{
get_years_lcs("it")
get_years_lcs(iso2 = c("de", "jp"))
}
