% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_and_years.R
\name{get_years_lis}
\alias{get_years_lis}
\title{Print all the existing years in LIS for a given country.}
\usage{
get_years_lis(iso2)
}
\arguments{
\item{iso2}{A character vector with valid iso2 codes of countries present in LIS.}
}
\value{
A list, made of numeric vectors. Each elements corresponds to a country in LIS.
}
\description{
Print all the existing years in LIS for a given country.
}
\examples{
get_years_lis("it")
get_years_lis(iso2 = c("de", "jp"))
}
