% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_and_years.R
\name{get_years_lws}
\alias{get_years_lws}
\title{Print all the existing years in LWS for a given country.}
\usage{
get_years_lws(iso2)
}
\arguments{
\item{iso2}{A character vector with valid iso2 codes of countries present in LWS.}
}
\value{
A list, made of numeric vectors. Each elements corresponds to a country in LWS.
}
\description{
Print all the existing years in LWS for a given country.
}
\examples{
get_years_lws("it")
get_years_lws(iso2 = c("de", "jp"))
}
