% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{has_lissy_attributes}
\alias{has_lissy_attributes}
\title{A list with lissy attributes}
\usage{
has_lissy_attributes(lissy_object)
}
\arguments{
\item{lissy_object}{A list of tibbles or a tibble.}
}
\value{
A boolean.
}
\description{
A list with lissy attributes
}
\keyword{internal}
