% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{implement_adjust_by_lisppp}
\alias{implement_adjust_by_lisppp}
\title{Applies the deflator adjustment to an aggregate}
\usage{
implement_adjust_by_lisppp(
  file,
  file_name,
  database,
  variable,
  income_variable = NULL,
  ppp_data = "lissyrtools"
)
}
\arguments{
\item{file}{A list of LIS or LWS file}

\item{file_name}{The name of the LIS or LWS file.}

\item{variable}{A string with the name of the variable to which the adjustment should be applied.}

\item{income_variable}{Defaults to NULL. Is only relevant for LWS files. If the file is
LWS and 'income_variable = TRUE', the function will retrieve the deflator for
the year in which the income data was collected. This might not be the same
as the year when wealth variables were collected.}

\item{ppp_data}{An optional file with the deflators. Should be in the same
format as the tibble in 'lissyrtools::deflators'. If "lissyrtools" (default) the deflators
are imported from the package internal data.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\details{
To be used inside transform_adjust_by_lisppp().
}
\keyword{internal}
