% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{implement_equivalise_oecd}
\alias{implement_equivalise_oecd}
\title{Apply equivalise with the OECD scale}
\usage{
implement_equivalise_oecd(
  file,
  file_name,
  variable,
  value_other_adults = 0.7,
  value_children = 0.5
)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{variable}{A string with the variable to which equivalisation should be applied.}

\item{value_other_adults}{A real number. Defaults to 0.7. The value assigned to the
second to last adults in the household.}

\item{value_children}{A real number. Defaults to 0.7. The value assigned to children
in the household.}
}
\value{
A a file with the equivalised variable.
}
\description{
\lifecycle{experimental}
Applies the equivalisation by number of  adults and children in a household.
}
\details{
To be used inside transform_equivalise_oecd().
}
\keyword{internal}
