% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{implement_false_zeros_to_na}
\alias{implement_false_zeros_to_na}
\title{Apply the recoding of zeros into missing values if all values are zero}
\usage{
implement_false_zeros_to_na(file, file_name, variable)
}
\arguments{
\item{file}{A tibble or data.frame with a LIS or LWS file.}

\item{file_name}{A string with the name of the LIS or LWS file.}

\item{variable}{A string with the variable to which top coding should be applied.}
}
\value{
A tibble containing the file with the recoded variable.
}
\description{
Lower-level function used within 'transform_false_zeros_to_na()' .
}
\keyword{internal}
