% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_subset.R
\name{implement_filter_age}
\alias{implement_filter_age}
\title{Apply 'transform_filter_age()' to a single file}
\usage{
implement_filter_age(file, file_name, from, to)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{from}{An integer with the lower boundary value (included in the sample).}

\item{to}{An integer with the higher boundary value (included in the sample).}
}
\value{
A a file with the filtered age variable.
}
\description{
\lifecycle{experimental}
Applies 'transform_filter_age()' to a single LIS/LWS file.
}
\details{
To be used inside 'transform_filter_age()' and 'transform_restrict_age()'.
}
\keyword{internal}
