% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{implement_negative_values_to_zero}
\alias{implement_negative_values_to_zero}
\title{Apply recoding negative values to zero}
\usage{
implement_negative_values_to_zero(file, file_name, variable)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file}

\item{variable}{A string with the variable to which recoding should be applied.}
}
\value{
A file with the recoded variable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Applies the recoding of negative values into zeroes in a variable of a single file.
}
\details{
To be used inside transform_negative_values_to_zero().
}
\keyword{internal}
