% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_subset.R
\name{implement_restrict_age}
\alias{implement_restrict_age}
\title{Apply 'transform_restrict_age()' to a single file}
\usage{
implement_restrict_age(file, file_name, variable, from, to)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{variable}{A character vector of length one with the variable that
will be recoded to NA when age is outside the boundaries (both included).}

\item{from}{An integer with the lower boundary value (included in the sample).}

\item{to}{An integer with the higher boundary value (included in the sample).}
}
\value{
A a file with the filtered age variable.
}
\description{
\lifecycle{experimental}
Applies 'transform_restrict_age()' to a single LIS/LWS file.
}
\keyword{internal}
