% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{implement_zeros_to_na}
\alias{implement_zeros_to_na}
\title{Apply recoding of zeros into missing values}
\usage{
implement_zeros_to_na(file, file_name, variable)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file}

\item{variable}{A string with the variable to which recoding should be applied.}
}
\value{
A file with the recoded variable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Applies the recoding of zeroes into missing values in a file for a single variable.
}
\details{
To be used inside transform_zeros_to_na().
}
\keyword{internal}
