% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{import_lisppp_data}
\alias{import_lisppp_data}
\title{Import lisppp dataset}
\usage{
import_lisppp_data(path_to_ppp_file = "lissy")
}
\arguments{
\item{path_to_ppp_file}{A character string indicating where the deflator values
can be found. If the value is 'lissyrtools', it will import the data from 'lissyrtools'.
Specifying 'lissy' (default) will read them from within the LISSY directory. Any other
value requires the full path specification to the deflators file.}
}
\value{
A tibble with the lisppp deflators.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Retrieves the lisppp deflators dataset.
}
\keyword{internal}
