% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sample_datasets.R
\name{import_sample_datasets_to_lissyuse}
\alias{import_sample_datasets_to_lissyuse}
\title{Import Sample Datasets to lissyuse}
\usage{
import_sample_datasets_to_lissyuse(data = NULL, database = "lis")
}
\arguments{
\item{data}{Optional. A named list of datasets to import. If \code{NULL}, all the default set is used.}

\item{database}{A character value. One of "lis" (default), "lws", or "lcs".}
}
\value{
A list.
}
\description{
Internal helper to import sample datasets into the \code{lissyuse} environment.
}
\keyword{internal}
