% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intervals_with_single_value}
\alias{intervals_with_single_value}
\title{Interval has single value}
\usage{
intervals_with_single_value(x)
}
\arguments{
\item{x}{string with the interval.}
}
\value{
A boolean indicating if the first and second element of the interval are
the same value.
}
\description{
Check if interval beginning and end are equal and thus it's composed
of a single value.
}
\details{
To be used inside 'compute_lorenz_curve()'
}
\keyword{internal}
