% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_iqr_top_bottom_coding.R
\name{iqr_top_bottom_coding}
\alias{iqr_top_bottom_coding}
\title{IQR-Based Top/Bottom Coding for a Single Variable}
\usage{
iqr_top_bottom_coding(
  x,
  w = NULL,
  times = 3,
  type = c("type_4", "type_2"),
  one_sided = NULL
)
}
\arguments{
\item{x}{Numeric vector. Variable to be top/bottom coded.}

\item{w}{Optional numeric vector of weights.}

\item{times}{Numeric. Multiplier for IQR to determine bounds (default is 3).}

\item{type}{Character. Quantile estimation type, passed to \code{compute_weighted_percentiles} function (e.g., \code{"type_4"}, \code{"type_2"}).}

\item{one_sided}{Character. \code{"top"}, \code{"bottom"}, or \code{NULL} (default) for two-sided coding.}
}
\value{
A numeric vector with values above or below the threshold replaced (on the original scale).
}
\description{
Internal utility that applies top and/or bottom coding to a numeric vector using an IQR rule
on the \code{log()} transformation.
}
\examples{
\dontrun{
data <- lissyrtools::lissyuse(data = "it16", vars = c("dhi"))
iqr_top_bottom_coding(data$it16h$dhi)
iqr_top_bottom_coding(data$it16h$dhi, data$it16h$hwgt, one_sided = "top")
}
}
