% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.all.same.value}
\alias{is.all.same.value}
\title{Test if all elements of a vector have the same value}
\usage{
is.all.same.value(x, na.rm = FALSE)
}
\description{
Test if all elements of a vector have the same value
}
\keyword{internal}
