% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.household.dataset}
\alias{is.household.dataset}
\title{Checks if the dataset is at household level}
\usage{
is.household.dataset(dataset_name)
}
\arguments{
\item{dataset_name}{The name of a LIS or LWS dataset.}
}
\value{
A boolean.
}
\description{
This function checks if there is an 'h' in the name of a LIS or LWS dataset.
}
\keyword{internal}
