% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_change_possible}
\alias{is_change_possible}
\title{Is possible to produce the desired format change}
\usage{
is_change_possible(current_format, to_format)
}
\arguments{
\item{current_format}{A string with the currant format.}

\item{to_format}{A string with the desired format.}
}
\value{
A boolean.
}
\description{
To be used inside 'change_file_name_format()'
}
\keyword{internal}
