% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_lissy_machine}
\alias{is_lissy_machine}
\title{Is a LISSY machine}
\usage{
is_lissy_machine()
}
\value{
a boolean.
}
\description{
Returns a boolean indicating if the code is running on a LISSY machine
}
\keyword{internal}
