% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{it14_h_lis}
\alias{it14_h_lis}
\title{Sample Household-Level LIS Dataset (1,000 Rows)}
\format{
Data frame with 1,000 rows and 112 columns.
}
\usage{
it14_h_lis
}
\description{
Sample Household-Level LIS Dataset (1,000 Rows)
}
\examples{
\dontrun{
it14_h_lis
}
}
\keyword{datasets}
