% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{it14_p_lws}
\alias{it14_p_lws}
\title{Sample Individual-Level LWS Dataset - Italy 2014}
\format{
Data frame with 2,328 rows and 133 columns.
}
\usage{
it14_p_lws
}
\description{
Sample Individual-Level LWS Dataset - Italy 2014
}
\examples{
\dontrun{
it14_p_lws
}
}
\keyword{datasets}
