% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{it16_h_lis}
\alias{it16_h_lis}
\title{Sample Household-Level LIS Dataset - Italy 2016}
\format{
Data frame with 1,000 rows and 112 columns.
}
\usage{
it16_h_lis
}
\description{
Sample Household-Level LIS Dataset - Italy 2016
}
\examples{
\dontrun{
it16_h_lis
}
}
\keyword{datasets}
