% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{it16_h_lws}
\alias{it16_h_lws}
\title{Sample Household-Level LWS Dataset (1,000 Rows)}
\format{
Data frame with 1,000 rows and 214 columns.
}
\usage{
it16_h_lws
}
\description{
Sample Household-Level LWS Dataset (1,000 Rows)
}
\examples{
\dontrun{
it16_h_lws
}
}
\keyword{datasets}
