% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{it16_p_lws}
\alias{it16_p_lws}
\title{Sample Individual-Level LWS Dataset - Italy 2016}
\format{
Data frame with 2,215 rows and 133 columns.
}
\usage{
it16_p_lws
}
\description{
Sample Individual-Level LWS Dataset - Italy 2016
}
\examples{
\dontrun{
it16_p_lws
}
}
\keyword{datasets}
