% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{it20_p_lis}
\alias{it20_p_lis}
\title{Sample Individual-Level LIS Dataset - Italy 2020}
\format{
Data frame with 1,000 rows and 97 columns.
}
\usage{
it20_p_lis
}
\description{
Sample Individual-Level LIS Dataset - Italy 2020
}
\examples{
\dontrun{
it20_p_lis
}
}
\keyword{datasets}
