% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy_14_lws_h}
\alias{italy_14_lws_h}
\title{Sample Household-Level LWS Dataset (1,000 Rows)}
\format{
Data frame with 1,000 rows and 214 columns.
}
\usage{
italy_14_lws_h
}
\description{
This dataset provides household-level data from the LWS database with 1,000 rows,
designed for exploration and user interaction. It is automatically loaded with
\code{lissyuse_locally()} whenever Italy is selected along with at least one household-level
variable (e.g., "region_c") and the \code{lws} argument is set to \code{TRUE}.
}
\details{
The primary purpose of this dataset, and the \code{lissyuse_locally()} function, is to allow
users to build and test code locally, simulating its execution before running it on LISSY
(the LIS user interface) using the \code{lissyuse()} function.
}
\examples{
\dontrun{
italy_14_lws_h
}
}
\keyword{datasets}
