% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_vars_household_lcs}
\alias{key_vars_household_lcs}
\title{Key Household Variables in LCS}
\format{
Character vector
}
\usage{
key_vars_household_lcs
}
\description{
A character vector of key household variables specific to LCS datasets.
}
\keyword{internal}
