% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_vars_household_lis}
\alias{key_vars_household_lis}
\title{Key Household Variables in LIS}
\format{
Character vector
}
\usage{
key_vars_household_lis
}
\description{
A character vector of key household variables specific to LIS datasets.
}
\keyword{internal}
