% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_vars_household_lws}
\alias{key_vars_household_lws}
\title{Key Household Variables in LWS}
\format{
Character vector
}
\usage{
key_vars_household_lws
}
\description{
A character vector of key household variables specific to LWS datasets.
}
\keyword{datasets}
