% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_vars_person_lis}
\alias{key_vars_person_lis}
\title{Key Person Variables in LIS}
\format{
Character vector
}
\usage{
key_vars_person_lis
}
\description{
A character vector of key person variables specific to LIS datasets.
}
\keyword{datasets}
