% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_both_hp_variables}
\alias{lcs_both_hp_variables}
\title{LCS Both Household and Person Variables}
\format{
Character vector
}
\usage{
lcs_both_hp_variables
}
\description{
A character vector combining household and person variables in LCS datasets.
}
\keyword{internal}
