% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_categorical_variables}
\alias{lcs_categorical_variables}
\title{LCS Categorical Variables}
\format{
Character vector
}
\usage{
lcs_categorical_variables
}
\description{
A character vector containing the names of categorical variables in LCS datasets.
}
\keyword{internal}
