% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_continuous_variables}
\alias{lcs_continuous_variables}
\title{LCS Continuous Variables}
\format{
Character vector
}
\usage{
lcs_continuous_variables
}
\description{
A character vector containing the names of continuous variables in LCS datasets.
}
\keyword{internal}
