% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_country_specific_variables}
\alias{lcs_country_specific_variables}
\title{LCS Country-Specific Variables}
\format{
Character vector
}
\usage{
lcs_country_specific_variables
}
\description{
A character vector of variables specific to certain countries in LCS datasets.
}
\keyword{internal}
