% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_household_variables}
\alias{lcs_household_variables}
\title{LCS Household Variables}
\format{
Character vector
}
\usage{
lcs_household_variables
}
\description{
A character vector containing the names of household-level variables in LCS datasets.
}
\keyword{internal}
