% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_id_variables}
\alias{lcs_id_variables}
\title{LCS ID Variables}
\format{
Character vector
}
\usage{
lcs_id_variables
}
\description{
A character vector containing the ID variables in LCS datasets.
}
\keyword{internal}
