% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_person_variables}
\alias{lcs_person_variables}
\title{LCS Person Variables}
\format{
Character vector
}
\usage{
lcs_person_variables
}
\description{
A character vector containing the names of person-level variables in LCS datasets.
}
\keyword{internal}
