% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_technical_variables}
\alias{lcs_technical_variables}
\title{LCS Technical Variables}
\format{
Character vector
}
\usage{
lcs_technical_variables
}
\description{
A character vector of technical variables used in LCS datasets.
}
\keyword{internal}
