% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_variables}
\alias{lcs_variables}
\title{LCS Variables}
\format{
A character vector of length 578.
}
\usage{
lcs_variables
}
\description{
A character vector containing the names of the variables available in LCS.
}
\keyword{datasets}
