% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcs_weight_variables}
\alias{lcs_weight_variables}
\title{LCS Weight Variables}
\format{
Character vector
}
\usage{
lcs_weight_variables
}
\description{
A character vector of weight variables in LCS datasets.
}
\keyword{internal}
