% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_both_hp_variables}
\alias{lis_both_hp_variables}
\title{LIS Both Household and Person Variables}
\format{
Character vector
}
\usage{
lis_both_hp_variables
}
\description{
A character vector combining both household and person variables in LIS datasets.
}
\keyword{internal}
