% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_categorical_variables}
\alias{lis_categorical_variables}
\title{LIS Categorical Variables}
\format{
Character vector
}
\usage{
lis_categorical_variables
}
\description{
A character vector containing the names of categorical variables in LIS datasets.
}
\keyword{datasets}
