% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_continuous_variables}
\alias{lis_continuous_variables}
\title{LIS Continuous Variables}
\format{
Character vector
}
\usage{
lis_continuous_variables
}
\description{
A character vector containing the names of continuous variables in LIS datasets.
}
\keyword{datasets}
