% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_country_specific_variables}
\alias{lis_country_specific_variables}
\title{LIS Country-Specific Variables}
\format{
Character vector
}
\usage{
lis_country_specific_variables
}
\description{
A character vector of variables specific to certain countries in LIS datasets.
}
\keyword{internal}
