% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_household_variables}
\alias{lis_household_variables}
\title{LIS Household Variables}
\format{
Character vector
}
\usage{
lis_household_variables
}
\description{
A character vector containing the names of household-level variables in LIS datasets.
}
\keyword{internal}
