% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_id_variables}
\alias{lis_id_variables}
\title{LIS ID Variables}
\format{
Character vector
}
\usage{
lis_id_variables
}
\description{
A character vector containing the ID variables in LIS datasets.
}
\keyword{datasets}
