% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_income_variables}
\alias{lis_income_variables}
\title{LIS Income Variables}
\format{
Character vector
}
\usage{
lis_income_variables
}
\description{
A character vector containing income-related variables in LIS datasets.
}
\keyword{datasets}
