% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_person_variables}
\alias{lis_person_variables}
\title{LIS Person Variables}
\format{
Character vector
}
\usage{
lis_person_variables
}
\description{
A character vector containing the names of person-level variables in LIS datasets.
}
\keyword{internal}
