% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_technical_variables}
\alias{lis_technical_variables}
\title{LIS Technical Variables}
\format{
Character vector
}
\usage{
lis_technical_variables
}
\description{
A character vector of technical variables used in LIS datasets.
}
\keyword{datasets}
