% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_variables}
\alias{lis_variables}
\title{LIS Variables}
\format{
A character vector of length 199.
}
\usage{
lis_variables
}
\description{
A character vector containing the names of the variables available in LIS.
}
\keyword{datasets}
