% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lis_weight_variables}
\alias{lis_weight_variables}
\title{LIS Weight Variables}
\format{
Character vector
}
\usage{
lis_weight_variables
}
\description{
A character vector of weight variables in LIS datasets.
}
\keyword{internal}
