% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lissyrtools_package_documentation.R
\docType{package}
\name{lissyrtools}
\alias{lissyrtools-package}
\alias{lissyrtools}
\title{lissyrtools: Functions that can be used within LISSY}
\description{
A package with the tools needed to develop scripts with LIS data.
It allows users to:
\itemize{
\item Read LIS data within the LISSY environment, or to use LIS sample files locally.
\item Carry out commonly performed data cleaning tasks.
\item Compute estimates from microdata.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lis-cross-national-data-center.github.io/lissyrtools/}
}

}
\author{
\strong{Maintainer}: Gonçalo Marques \email{marques@lisdatacenter.org}

Authors:
\itemize{
  \item Josep Espasa Reig \email{jespasareig@gmail.com}
}

}
