% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lissyuse.r
\name{load_datasets}
\alias{load_datasets}
\title{Load Dataset Names Based on Criteria}
\usage{
load_datasets(data = NULL, database = "lis", from = NULL, to = NULL)
}
\arguments{
\item{data}{Optional character vector specifying ccyy pairs or iso2 codes.}

\item{database}{A character value. One of "lis" (default), "lws", or "lcs".}

\item{from}{Optional numeric lower bound for dataset years.}

\item{to}{Optional numeric upper bound for dataset years.}
}
\value{
A character vector of dataset names matching the criteria.
}
\description{
Internal function to filter and return dataset names based on input criteria such as
specified iso2 codes or ccyy pairs, database type (LIS, LWS or LCS), and optional year range.
}
\keyword{internal}
