% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lws_both_hp_variables}
\alias{lws_both_hp_variables}
\title{LWS Both Household and Person Variables}
\format{
Character vector
}
\usage{
lws_both_hp_variables
}
\description{
A character vector combining household and person variables in LWS datasets.
}
\keyword{datasets}
