% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lws_household_variables}
\alias{lws_household_variables}
\title{LWS Household Variables}
\format{
Character vector
}
\usage{
lws_household_variables
}
\description{
A character vector containing household-level variables in LWS datasets.
}
\keyword{datasets}
