% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lws_id_variables}
\alias{lws_id_variables}
\title{LWS ID Variables}
\format{
Character vector
}
\usage{
lws_id_variables
}
\description{
A character vector containing ID variables in LWS datasets.
}
\keyword{internal}
