% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lws_person_variables}
\alias{lws_person_variables}
\title{LWS Person Variables}
\format{
Character vector
}
\usage{
lws_person_variables
}
\description{
A character vector containing person-level variables in LWS datasets.
}
\keyword{internal}
