% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lws_variables}
\alias{lws_variables}
\title{LWS Variables}
\format{
A character vector of length 336.
}
\usage{
lws_variables
}
\description{
A character vector containing the names of the variables available in LWS.
}
\keyword{datasets}
