% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lws_wealth_categorical_variables}
\alias{lws_wealth_categorical_variables}
\title{LWS Wealth Categorical Variables}
\format{
Character vector
}
\usage{
lws_wealth_categorical_variables
}
\description{
A character vector of categorical wealth-related variables in LWS datasets.
}
\keyword{datasets}
