% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lws_wealth_continuous_variables}
\alias{lws_wealth_continuous_variables}
\title{LWS Wealth Continuous Variables}
\format{
Character vector
}
\usage{
lws_wealth_continuous_variables
}
\description{
A character vector of continuous wealth-related variables in LWS datasets.
}
\keyword{datasets}
