% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lws_wealth_country_specific_variables}
\alias{lws_wealth_country_specific_variables}
\title{LWS Wealth Country-Specific Variables}
\format{
Character vector
}
\usage{
lws_wealth_country_specific_variables
}
\description{
A character vector of country-specific wealth variables in LWS datasets.
}
\keyword{internal}
