% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dataset_levels.R
\name{merge_dataset_levels}
\alias{merge_dataset_levels}
\title{Merge household and person-level files}
\usage{
merge_dataset_levels(lissy_hfiles, lissy_pfiles)
}
\arguments{
\item{lissy_hfiles}{A list with household-level files.}

\item{lissy_pfiles}{A list with person-level files.}
}
\value{
A list with the merged files.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Merges the household and person-level files of datasets.

The level of the files in 'lissy_hfiles' and 'lissy_pfiles' is determined by
the 'level' attribute of the files.
}
