% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dataset_levels.R
\name{merge_lis_dataset_levels}
\alias{merge_lis_dataset_levels}
\alias{merge_lws_dataset_levels}
\title{Merges a hfile with its pfile}
\usage{
merge_lis_dataset_levels(hfile, pfile, hfile_name, pfile_name)

merge_lws_dataset_levels(hfile, pfile, hfile_name, pfile_name)
}
\description{
To be used only within 'merge_dataset_levels()'
}
\keyword{internal}
