% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metis_countries_df}
\alias{metis_countries_df}
\title{Countries Names and ISO Codes.}
\format{
A data frame with 3 variables:
}
\usage{
metis_countries_df
}
\description{
A data frame containing the country names, and its respective iso2 and iso3 codes.
}
\keyword{datasets}
