% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_or_zero_vars_all}
\alias{missing_or_zero_vars_all}
\title{Variable Status Data Frame}
\format{
A data frame with 7 variables:
}
\usage{
missing_or_zero_vars_all
}
\description{
A data frame that identifies whether a variable consists only of zeros or missing values for a given country and year.
}
\keyword{datasets}
