% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_options.R
\name{missing_values_in_variable_warning}
\alias{missing_values_in_variable_warning}
\title{Missing values in variable warning}
\usage{
missing_values_in_variable_warning(file, file_name, variable)
}
\arguments{
\item{file}{A LIS or LWS file.}

\item{file_name}{The name of the LIS or LWS file.}

\item{variable}{A string with the variable name.}
}
\description{
Triggers a warning if there are missing values in 'variable' and the
warning has not been triggered before.
}
\details{
Uses an option to store the names of the files and variables with missing values
that have already triggered a warning before. The option is stored as:
'{file_name}\emph{warning_NAs}{variable}'.
}
\keyword{internal}
