% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mx14_p_lis}
\alias{mx14_p_lis}
\title{Sample Individual-Level LIS Dataset - Mexico 2014}
\format{
Data frame with 1,000 rows and 97 columns.
}
\usage{
mx14_p_lis
}
\description{
Sample Individual-Level LIS Dataset - Mexico 2014
}
\examples{
\dontrun{
mx14_p_lis
}
}
\keyword{datasets}
