% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mx16_p_lcs}
\alias{mx16_p_lcs}
\title{Sample Individual-Level LCS Dataset - MX 2016}
\format{
Data frame with 3,648 rows and 105 columns.
}
\usage{
mx16_p_lcs
}
\description{
Sample Individual-Level LCS Dataset - MX 2016
}
\examples{
\dontrun{
mx16_p_lcs
}
}
\keyword{datasets}
