% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mx16_p_lis}
\alias{mx16_p_lis}
\title{Sample Individual-Level LIS Dataset (1,000 Rows)}
\format{
Data frame with 1,000 rows and 97 columns.
}
\usage{
mx16_p_lis
}
\description{
Sample Individual-Level LIS Dataset (1,000 Rows)
}
\examples{
\dontrun{
mx16_p_lis
}
}
\keyword{datasets}
