% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mx18_h_lis}
\alias{mx18_h_lis}
\title{Sample Household-Level LIS Dataset - Mexico 2018}
\format{
Data frame with 1,000 rows and 112 columns.
}
\usage{
mx18_h_lis
}
\description{
Sample Household-Level LIS Dataset - Mexico 2018
}
\examples{
\dontrun{
mx18_h_lis
}
}
\keyword{datasets}
