% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mx20_h_lcs}
\alias{mx20_h_lcs}
\title{Sample Household-Level LCS Dataset - MX 2020}
\format{
Data frame with 1,000 rows and 552 columns.
}
\usage{
mx20_h_lcs
}
\description{
Sample Household-Level LCS Dataset - MX 2020
}
\examples{
\dontrun{
mx20_h_lcs
}
}
\keyword{datasets}
