% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mx20_p_lcs}
\alias{mx20_p_lcs}
\title{Sample Individual-Level LCS Dataset - MX 2020}
\format{
Data frame with 3,447 rows and 105 columns.
}
\usage{
mx20_p_lcs
}
\description{
Sample Individual-Level LCS Dataset - MX 2020
}
\examples{
\dontrun{
mx20_p_lcs
}
}
\keyword{datasets}
