% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_numbers_smaller}
\alias{n_numbers_smaller}
\title{Number of elements smaller than input}
\usage{
n_numbers_smaller(x, numbers)
}
\arguments{
\item{x}{A numeric vector with input elements}

\item{numbers}{A numeric vector with numbers which}
}
\value{
A vector of integer values of same length as x
}
\description{
Computes the number of elements in 'numbers' smaller than each of the
elements in input 'x'.
}
\keyword{internal}
