% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{paste_lissy_attributes}
\alias{paste_lissy_attributes}
\title{Paste lissy attributes}
\usage{
paste_lissy_attributes(lissy_object, lissy_attributes)
}
\arguments{
\item{lissy_object}{A list of tibbles or a tibble.}

\item{lissy_attributes}{A list of attributes or a NULL value, as obtained from 'get_lissy_attributes()'.}
}
\value{
A tibble.
}
\description{
Implements returns the 'lissy_object' with the 'lissy_attributes' if
'lissy_attributes is not NULL. Else, it just returns 'lissy_object'.
}
\keyword{internal}
