% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot an indicator using a bar plot.}
\usage{
plot_bar(results_df)
}
\arguments{
\item{results_df}{A data frame with the results of the indicator.}
}
\value{
A ggplot2 plot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
