% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lorenz_curve}
\alias{plot_lorenz_curve}
\title{Plot Lorenz Curve.}
\usage{
plot_lorenz_curve(list_datasets, variable, plot_theme = NULL, na.rm = FALSE)
}
\arguments{
\item{list_datasets}{A list of LIS or LWS datasets.}

\item{variable}{A character vector of length one.}
}
\description{
\lifecycle{experimental}
Computes the equivalised percentiles of an indicator and plots the Lorenz
Curve for one or multiple LIS or LWS datasets into a single graph.
}
\examples{
\dontrun{
datasets <- read_lissy_files(c("fr84h", "fr94h", "fr10h"))
plot_lorenz_curve(datasets, "dhi")
}
}
