% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_atkinson}
\alias{print_atkinson}
\title{Print the Atkinson Index}
\usage{
print_atkinson(
  lissy_files,
  variable,
  epsilon,
  weight = NULL,
  na.rm = FALSE,
  files_level = NULL,
  variable_level = NULL
)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character string indicating the variable for which the Atkinson index needs to be computed.}

\item{epsilon}{A numeric vector of length one. The inequality aversion parameter. Needs to be epsilon > 0.}

\item{weight}{A string with the name of the variable in 'file' that should be used as sample weights.
If NULL (default), the function tries to guess the needed weight to compute the Atkinson index.
This guess is made based on the information from files_level and variable_level.}

\item{na.rm}{A boolean. Indicates if NAs should be ignored. Defaults to FALSE.}

\item{files_level}{A string indicating the level of the file. Valid inputs are:
'household', 'h', 'person' or 'p'. If NULL (default), the file level will be retrieved from the 'lissy_files' attributes.}

\item{variable_level}{Level of the variable. Should be either 'household', 'h', 'person' or 'p'.
If NULL (default), the function will try to guess the level of the variable.
This is done by comparing the value in 'variable' with pre-set lists of variables.}
}
\value{
A numeric vector containing the Atkinson index for each file.
}
\description{
\lifecycle{experimental}
Computes and displays the Atkinson index for a given variable across multiple files.
}
\examples{
\dontrun{
lissy_files <- read_lissy_files(c("file1", "file2"))
print_atkinson(lissy_files = lissy_files, variable = "income", epsilon = 0.5)
}
}
