% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_gini}
\alias{print_gini}
\title{Print the Gini Coefficient}
\usage{
print_gini(
  lissy_files,
  variable,
  weight = NULL,
  na.rm = FALSE,
  files_level = NULL,
  variable_level = NULL
)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character string indicating the variable for which the Gini coefficient needs to be computed.}

\item{weight}{A string with the name of the variable in 'file' that should be used as sample weights.
If NULL (default), the function tries to guess the needed weight to compute the Gini coefficient.
This guess is made based on the information from files_level and variable_level.}

\item{na.rm}{A boolean. Indicates if NAs should be ignored. Defaults to FALSE.}

\item{files_level}{A string indicating the level of the file. Valid inputs are:
'household', 'h', 'person' or 'p'. If NULL (default), the file level will be retrieved from the 'lissy_files' attributes.}

\item{variable_level}{Level of the variable. Should be either 'household', 'h', 'person' or 'p'.
If NULL (default), the function will try to guess the level of the variable.
This is done by comparing the value in 'variable' with pre-set lists of variables.}
}
\value{
A numeric vector containing the Gini coefficient for each file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Computes and displays the Gini coefficient for a given variable across multiple files.
}
\examples{
\dontrun{
lissy_files <- read_lissy_files(c("file1", "file2"))
print_gini(lissy_files = lissy_files, variable = "income")
}
}
