% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_percentiles}
\alias{print_percentiles}
\title{Print percentiles.}
\usage{
print_percentiles(
  lissy_files,
  variable,
  breaks = seq(0, 1, 0.1),
  weight = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character vector of length one.}

\item{breaks}{A numeric vector with specifying the percentiles that should be computed. Defaults to deciles.}

\item{weight}{A string with the name of the variable in 'file' that should be used as sample weights.}

\item{na.rm}{A boolean indicating if missing values should be ignored. Defaults to FALSE.}
}
\value{
A tibble with percentile absolute and cummulative values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Computes and displays the percentiles and cumulative percentiles of a
variable.
}
\examples{
\dontrun{
lissy_files <- read_lissy_files(c("fr84h", "fr94h", "fr10h"))
print_percentiles(lissy_files = lissy_files, variable = "dhi")
}
}
