% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_ratio}
\alias{print_ratio}
\title{Print the Ratio Index}
\usage{
print_ratio(
  lissy_files,
  variable,
  ratio,
  weight = NULL,
  na.rm = FALSE,
  files_level = NULL,
  variable_level = NULL
)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character string indicating the variable for which the ratio index needs to be computed.}

\item{ratio}{A vector of two numeric values between 0 and 1.
Defines the percentiles in the numerator and denominator respectively.
E.g. (0.9, 0.1) computes the 90/10 ratio.}

\item{weight}{A string with the name of the variable in 'file' that should be used as sample weights.
If NULL (default), the function tries to guess the needed weight to compute the ratio index.
This guess is made based on the information from files_level and variable_level.}

\item{na.rm}{A boolean. Indicates if NAs should be ignored. Defaults to FALSE.}

\item{files_level}{A string indicating the level of the file. Valid inputs are:
'household', 'h', 'person' or 'p'. If NULL (default), the file level will be retrieved from the 'lissy_files' attributes.}

\item{variable_level}{Level of the variable. Should be either 'household', 'h', 'person' or 'p'.
If NULL (default), the function will try to guess the level of the variable.
This is done by comparing the value in 'variable' with pre-set lists of variables.}
}
\value{
A numeric vector containing the ratio index for each file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Computes and displays the ratio index for a given variable across multiple files.
}
\examples{
\dontrun{
lissy_files <- read_lissy_files(c("file1", "file2"))
print_ratio(lissy_files = lissy_files, variable = "income", ratio = c(0.9, 0.1))
}
}
