% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_file_name_format}
\alias{read_file_name_format}
\title{Read the format of a file name}
\usage{
read_file_name_format(file_name)
}
\arguments{
\item{file_name}{A string with the name of the file.}
}
\value{
A string indicating the format of the file.
}
\description{
Read the format of a file name
}
\keyword{internal}
