% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lissy_files.R
\name{read_lissy_files}
\alias{read_lissy_files}
\title{Read files in LISSY}
\usage{
read_lissy_files(files, col_select = NULL, full_year_names = TRUE)
}
\arguments{
\item{files}{A character vector containing file names in 'ccyyl' format.
All files need to be at the same level (i.e. 'household', 'person').
E.g. c("it14h", "us16h")}

\item{col_select}{A character vector with the name of the variables
which should be selected from the files. E.g. c("hid", "dhi", "hifactor").}

\item{full_year_names}{A boolean. Should the name of the imported file be changed to 'ccyyyyl' format.}
}
\value{
A named list with the loaded files.
}
\description{
\lifecycle{experimental}
Reads multiple LIS, LWS or ERFLIS files in the LISSY interface.
}
\details{
All files need to be from the same database (e.g. LIS, LWS) and level
(e.g. 'household', 'person') Computes the full year with four digits by default.
}
\examples{
\dontrun{
read_lissy_files(c("fr84h", "fr94h", "fr10h"))
}
}
