% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lissy_files.R
\name{read_lissy_files_locally}
\alias{read_lissy_files_locally}
\title{Read files locally}
\usage{
read_lissy_files_locally(
  files,
  path_to_files,
  col_select = NULL,
  full_year_names = TRUE
)
}
\arguments{
\item{files}{A character vector containing file names. These need to contain
the country, year, database and level elements at the beginning of their name in 'ccyydl', or 'ccyyyydl' format.
E.g. c("it14ih", "us16ih") or c("it14ih_modified.dta", "us16ih_mydata.dta")}

\item{path_to_files}{A character string with the directory in which the files can be found.
The order should match that of the files passed in 'files'.}

\item{col_select}{A character vector with the name of the variables
which should be selected from the files. E.g. c("hid", "dhi", "hifactor").}

\item{full_year_names}{A boolean. Should the name of the imported file be changed to 'ccyyyyl' format.}
}
\value{
A named list with the loaded files.
}
\description{
\lifecycle{deprecated}
Reads multiple LIS or LWS files outside of the LISSY interface.
}
\examples{
\dontrun{
read_lissy_files_locally(c("fr84h", "fr94h", "fr10h"), path_to_files = "/home/user/files/")
}
}
