% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lissy_files.R
\name{read_rename_files}
\alias{read_rename_files}
\title{Rename read files}
\usage{
read_rename_files(list_files, new_names)
}
\arguments{
\item{list_files}{A list of LIS or LWS datasets.}

\item{new_names}{A character string with the new names of files.}
}
\value{
A list of loaded files with the new names.
}
\description{
'
\lifecycle{experimental}
Uses a character vector to manually change the names files that have been previously read.
}
\examples{
\dontrun{
files <- read_lissy_files_locally(c("fr84h", "fr94h", "fr10h"))
files \%<>\%
    read_rename_files(new_names = c("France_1984_household",
                                    "France_1994_household",
                                    "France_2010_household"))
}
}
