% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_canada_lws_missing_weights_in_p_file}
\alias{remove_canada_lws_missing_weights_in_p_file}
\title{Is Canada LWS on the loaded data ?}
\usage{
remove_canada_lws_missing_weights_in_p_file(list, wgt_name)
}
\value{
a list
}
\description{
Auxilliary function to remove any p-level Canadian datasets in LWS, from the list of loaded datasets, as it has no weights
}
\keyword{internal}
