% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_dname_with_missings_in_weights}
\alias{remove_dname_with_missings_in_weights}
\title{Is Canada LWS on the loaded data ?}
\usage{
remove_dname_with_missings_in_weights(list, wgt_name)
}
\value{
a list
}
\description{
Auxilliary function to remove any p-level Canadian datasets in LWS, from the list of loaded datasets, as it has no weights
}
\keyword{internal}
