% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lissy_files.R
\name{retrieve_database_from_directory}
\alias{retrieve_database_from_directory}
\title{Get the database from a directory string}
\usage{
retrieve_database_from_directory()
}
\description{
Get the database from a directory string
}
\keyword{internal}
