% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lissy_files.R
\name{retrieve_names_key_variables}
\alias{retrieve_names_key_variables}
\title{Retrieve names of key variables}
\usage{
retrieve_names_key_variables(database, level)
}
\arguments{
\item{database}{A string indicating the database of the key variables to be
retrieved (e.g. 'lis', 'i', 'lws', 'w')}

\item{level}{A string indicating the level of the key variables to be
retrieved (e.g. 'person', 'p', 'household', 'h')}
}
\value{
A character vector with the names of key variables for a given
database and level.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Retrieve the names of key variables for a given database and level. To be
used only within 'read_lissy_files()' and 'read_lissy_files_locally()'.
}
\keyword{internal}
