% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_desired_format}
\alias{return_desired_format}
\title{Return a file name with the requested format}
\usage{
return_desired_format(country, year, database, level, to_format)
}
\arguments{
\item{country}{A string with the country as 'cc'.}

\item{year}{A string with four digit year as 'yyyy'.}

\item{database}{A string with database as 'd'.}

\item{level}{A string with level as 'l'.}

\item{to_format}{A string with the desired format.}
}
\value{
A string with the file name in the stipulated format.
}
\description{
Combines the information passed and returns it in the format stipulated in
argument 'to_format'.
}
\details{
To be used inside 'change_file_name_format()'.
}
\keyword{internal}
