% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_gini.R
\name{run_weighted_gini}
\alias{run_weighted_gini}
\title{Compute Weighted Gini Index Across a List of Data Frames}
\usage{
run_weighted_gini(data_list, var_name, wgt_name = NULL, na.rm = TRUE)
}
\arguments{
\item{data_list}{A named list of data frames, (e.g., across countries or years).}

\item{var_name}{A string specifying the variable name (e.g., "dhi", "pilabour") to compute the Gini index on.}

\item{wgt_name}{An optional string specifying the weight variable to be used. If \code{NULL}, equal weights are assumed.}

\item{na.rm}{Logical. If \code{TRUE}, missing values in \code{var_name} and \code{wgt_name} are removed.}
}
\value{
A named list. Each list element is named by country and contains a named numeric vector, where the names are years and the values are the computed statistics.
}
\description{
Compute Weighted Gini Index Across a List of Data Frames
}
\examples{
\dontrun{
library(lissyrtools)
library(purrr)

datasets <- lissyrtools::lissyuse(data = c("de", "es", "uk"), vars = c("dhi"), from = 2016)

datasets \%>\% 
 map(~ .x \%>\% mutate(new_wgt = hwgt * nhhmem)) \%>\%  
 apply_iqr_top_bottom_coding("dhi", "hwgt") \%>\%  
 apply_sqrt_equivalisation("dhi") \%>\% 
 run_weighted_gini("dhi", "new_wgt")
}
}
