% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean.R
\name{run_weighted_mean}
\alias{run_weighted_mean}
\title{Compute Weighted Mean Across a List of Data Frames (with optional grouping)}
\usage{
run_weighted_mean(
  data_list,
  var_name,
  wgt_name = NULL,
  na.rm = TRUE,
  by = NULL
)
}
\arguments{
\item{data_list}{A named list of data frames, (e.g., across countries or years).}

\item{var_name}{A string specifying the variable name (e.g., "dhi", "pilabour") to compute the mean on.}

\item{wgt_name}{An optional string specifying the weight variable to be used. If \code{NULL}, equal weights are assumed.}

\item{na.rm}{Logical. If \code{TRUE}, missing values in \code{var_name} or \code{wgt_name} are removed.}

\item{by}{Optional string giving the name of a categorical variable to split the data within each data frame before computing the mean.}
}
\value{
A named list of data frames.
}
\description{
Compute Weighted Mean Across a List of Data Frames (with optional grouping)
}
\examples{
\dontrun{
data <- lissyrtools::lissyuse(data = c("de", "es", "uk"), vars = c("dhi", "age", "pi11", "region_c", "area_c", "educ", "emp"), from = 2016)

data \%>\% 
purrr::map(~ .x \%>\% filter(relation == 1000)) \%>\%
run_weighted_mean("dhi", "hpopwgt")

data \%>\% 
purrr::map(~ .x \%>\% filter(age > 25 & age <65)) \%>\%
run_weighted_mean("pi11", "ppopwgt", by = "educ")

}
}
