% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_poverty.R
\name{run_weighted_poverty_shortfall}
\alias{run_weighted_poverty_shortfall}
\title{Compute the Weighted Poverty Gap (Shortfall)}
\usage{
run_weighted_poverty_shortfall(
  data_list,
  var_name,
  wgt_name = NULL,
  times_median = 0.5,
  daily_poverty_line = NULL,
  type = c("type_4", "type_2"),
  percent = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data_list}{A named list of data frames.}

\item{var_name}{A string specifying the variable name (e.g., \code{"dhi"}, \code{"pilabour"}).}

\item{wgt_name}{A string (optional). The name of the weight variable. If \code{NULL}, equal weights are assumed.}

\item{times_median}{A numeric scalar. The multiple of the median used to define the poverty threshold (default is \code{0.5}).}

\item{daily_poverty_line}{A numeric scalar representing the \strong{absolute poverty threshold per day}. (default is \code{NULL}).}

\item{type}{A character vector indicating the percentile estimation type (passed to \code{compute_weighted_percentiles}). Default is \code{"type_4"}.}

\item{percent}{Logical. If \code{TRUE}, returns the relative shortfall as a percentage of the poverty line. If \code{FALSE}, returns the absolute daily shortfall (default is \code{FALSE}).}

\item{na.rm}{Logical. Should missing values be removed before computation? Default is \code{TRUE}.}
}
\value{
A named list. Each list element is named by country and contains a named numeric vector, where the names are years and the values represent:
\itemize{
\item The average daily shortfall in monetary units (if \code{percent = FALSE}), or
\item The average relative shortfall in percentage terms (if \code{percent = TRUE}), among individuals or households below the poverty threshold.
}
}
\description{
Calculates the average shortfall (gap) between individual/household income and the relative poverty line, weighted by the population weights and restricted to those below the poverty line.
}
\details{
When \code{daily_poverty_line} is specified, the function assumes the threshold is expressed in daily units and converts it to an annual threshold by multiplying it by 365.
It is \strong{critical} that the daily poverty threshold be expressed in the \strong{same monetary unit} as the target income variable (\code{var_name}). For example, if income is reported in PPP-adjusted USD, the threshold must also be in PPP-adjusted USD. The function does not perform currency or PPP conversions.
}
\examples{
\dontrun{
library(lissyrtools)
library(purrr)
library(dplyr)

datasets <- lissyrtools::lissyuse(data = c("de", "es", "uk"), vars = c("dhi"), from = 2016)

datasets <- datasets \%>\%
  map(~ .x \%>\% filter(!is.na(dhi))) \%>\%
  map(~ .x \%>\% mutate(new_wgt = hwgt * nhhmem)) \%>\%
  apply_iqr_top_bottom_coding("dhi", "hwgt", type = "type_2") \%>\%
  apply_sqrt_equivalisation("dhi")

# Compute average poverty shortfall in daily monetary terms
avg_gap <- run_weighted_poverty_shortfall(
  data_list = datasets,
  var_name = "dhi",
  wgt_name = "new_wgt"
)

# Compute average poverty shortfall in percentage of the poverty line
percentage_gap <- run_weighted_poverty_shortfall(
  data_list = datasets,
  var_name = "dhi",
  wgt_name = "new_wgt",
  percent = TRUE
)

# Compute average poverty shortfall in daily monetary terms, converted to international dollars at 2017 prices. 
avg_gap_dollars <- datasets \%>\% 
  apply_ppp_adjustment("dhi", "lis", "lisppp") \%>\% 
  run_weighted_poverty_shortfall(
  var_name = "dhi",
  wgt_name = "new_wgt"
)

# Compute average poverty shortfall in daily monetary terms, from a specified poverty threshold in international dollars at 2017 prices. 
avg_gap_absolute_line <- datasets \%>\% 
  apply_ppp_adjustment("dhi", "lis", "lisppp") \%>\% 
  run_weighted_poverty_shortfall(
  var_name = "dhi",
  wgt_name = "new_wgt", 
  daily_poverty_line  = 10
)


print(avg_gap)
print(percentage_gap)
print(avg_gap_dollars)
print(avg_gap_absolute_line)
}
}
