% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ratios.R
\name{run_weighted_ratios}
\alias{run_weighted_ratios}
\title{Compute Percentile Ratios}
\usage{
run_weighted_ratios(
  data_list,
  var_name,
  wgt_name = NULL,
  upper_percentile,
  lower_percentile,
  type = c("type_4", "type_2"),
  na.rm = TRUE
)
}
\arguments{
\item{data_list}{A named list of data frames.}

\item{var_name}{A string. The name of the variable to analyze (e.g., "dhi", "income").}

\item{wgt_name}{A string (optional). The name of the weight variable. If \code{NULL}, equal weights are assumed.}

\item{upper_percentile}{A numeric scalar (between 0 and 1). The higher percentile (e.g., 0.9).}

\item{lower_percentile}{A numeric scalar (between 0 and 1). The lower percentile (e.g., 0.1).}

\item{type}{A character string indicating the percentile type used in \code{compute_weighted_percentiles}. Defaults to \code{"type_4"}.}

\item{na.rm}{Logical. Should missing values be removed before computation? Default is \code{TRUE}.}
}
\value{
A named list. Each element is named by country and contains a named numeric vector, where names are years and values are the percentile ratios.
}
\description{
This function calculates the ratio between two percentiles (e.g., P90/P10) for each dataset in a list.
}
\details{
If \code{upper_percentile} is less than \code{lower_percentile}, the values are automatically swapped and a warning is issued.
}
\examples{
\dontrun{
library(lissyrtools)
library(purrr)

datasets <- lissyrtools::lissyuse(data = c("de", "es", "uk"), vars = c("dhi"), from = 2016)

datasets <- datasets \%>\% 
 map(~ .x \%>\% filter(!is.na(dhi))) \%>\%
 map(~ .x \%>\% mutate(new_wgt = hwgt * nhhmem)) \%>\%
 apply_iqr_top_bottom_coding("dhi", "hwgt", type = "type_2") \%>\%
 apply_sqrt_equivalisation("dhi")


# Compute dhi  ratios

p90_p10 <- run_weighted_ratios(datasets, var_name = "dhi", wgt_name = "hwgt", 
                                upper_percentile = 0.9, lower_percentile = 0.1)
                                
p90_p50 <- run_weighted_ratios(datasets, var_name = "dhi", wgt_name = "hwgt", 
                                upper_percentile = 0.9, lower_percentile = 0.5) 
                                                                
p80_p20 <- run_weighted_ratios(datasets, var_name = "dhi", wgt_name = "hwgt", 
                                upper_percentile = 0.8, lower_percentile = 0.2) 
print(p90_p10)
print(p90_p50) 
print(p80_p20)
}
}
