% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_lissy_attributes}
\alias{set_lissy_attributes}
\title{Set lissy attributes}
\usage{
set_lissy_attributes(lissy_object, lissy_attributes)
}
\arguments{
\item{lissy_object}{A list of tibbles or a tibble.}

\item{lissy_attributes}{A list with three elements: 'level', 'merged_levels'
and 'database'.}
}
\value{
The same type of object as 'lissy_object'.
}
\description{
Sets the 'level', 'merged_levels' and 'database' attributes of an object equal to
the specified ones.
}
\keyword{internal}
