% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_and_years.R
\name{show_country_years_lis}
\alias{show_country_years_lis}
\title{Print all the existing years in LIS for a given country.}
\usage{
show_country_years_lis(iso2 = NULL)
}
\arguments{
\item{iso2}{A string with 2 characters, specifically an iso2 code present in show_countries_lis().}
}
\value{
A numeric vector.
}
\description{
Print all the existing years in LIS for a given country.
}
\examples{
show_country_years_lis("it")
show_country_years_lis("us")
}
