% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dataset_levels.R
\name{sort_lissy_list}
\alias{sort_lissy_list}
\title{Sort a list of lissy files}
\usage{
sort_lissy_list(list_to_sort, according_to)
}
\arguments{
\item{list_to_sort}{A list with pfiles or hfiles. Names of these files should be in 'ccyyyydl' format.}

\item{according_to}{A character vector with names on which the list should be sorted.
These can be in any format that can be transformed to 'ccyyyyd'.}
}
\description{
To be used only within 'merge_dataset_levels()'
}
\details{
Sorts a list of lissy household or person-level files according to the
names passed in argument 'according_to'.
}
\keyword{internal}
