% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_consecutive_values}
\alias{split_consecutive_values}
\title{Split vector into consecutive values}
\usage{
split_consecutive_values(x)
}
\arguments{
\item{x}{a vector with values that should be split into groups.}
}
\description{
Split a vector into groups of consecutive values.
}
\keyword{internal}
