% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lissyuse.r
\name{subset_datasets}
\alias{subset_datasets}
\title{Subset Datasets Based on Expression}
\usage{
subset_datasets(intermediate_data_to_filter, database = "lis", subset_expr)
}
\arguments{
\item{intermediate_data_to_filter}{List of data frames to subset.}

\item{database}{A character value. One of "lis" (default), "lws", or "lcs".}

\item{subset_expr}{An expression used to filter the datasets.}
}
\value{
List. A subsetted version of the input list based on the provided expression.
}
\description{
Internal function that subsets the given datasets based on a filtering expression.
Supports both LWS and LIS datasets.
}
\keyword{internal}
