% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_from_other_packages.R
\name{testDateTime}
\alias{testDateTime}
\title{Test if a value is a date-time object}
\usage{
testDateTime(x, what = c("either", "both", "timeVaries"))
}
\description{
Function from: Frank E Harrell Jr, with contributions from Charles Dupont and many others. (2020). Hmisc: Harrell Miscellaneous. R package version 4.4-1.
https://CRAN.R-project.org/package=Hmisc
}
\keyword{internal}
