% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_equivalise}
\alias{transform_equivalise}
\title{Equivalise by number of household members.}
\usage{
transform_equivalise(lissy_files, variable, eq_scale = 0.5)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files}

\item{variable}{A character vector of length one with the indicator that needs to be transformed.}

\item{eq_scale}{A real number.}
}
\value{
A list of tibbles with the transformed variable.
}
\description{
\lifecycle{experimental}
Scales a variable dividing it by 'nhhmem'^eq_scale. Where 'nhhmem' is the LIS
or LWS variable measuring the number of members in the household.
}
\details{
Throws a warning if the function is applied to a person-level variable.
}
