% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_false_zeros_to_na}
\alias{transform_false_zeros_to_na}
\title{Recode zeros into missing values if all values are zero}
\usage{
transform_false_zeros_to_na(lissy_files, variable)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character string with the name of the variable that should be adjusted.}
}
\value{
A list of tibbles with the adjusted variable.
}
\description{
Recodes all zeros in the selected variable into missing values (NAs) if (and
only if) all values are zeros.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Some LIS/LWS datasets have variables with only '0' values. These do not
represent 0s but NAs. This function transform the 0s into NAs if it finds
that all values are 0s and there are no other valid values.
}
