% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_negative_values_to_zero}
\alias{transform_negative_values_to_zero}
\title{Recode negative values to zero}
\usage{
transform_negative_values_to_zero(lissy_files, variable)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character string with the name of the variable that should be transformed.}
}
\value{
A list of tibbles with the recoded variable.
}
\description{
\lifecycle{experimental}
Recodes all negative values to zero for all files in a list.
}
\examples{
\dontrun{
library(dplyr)
library(magrittr)
lissy_files <- read_lissy_files(c("fr84h", "fr94h", "fr10h"))
lissy_files \%<>\%
    transform_negative_values_to_zero(variable = "dhi")
    }
}
