% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_subset.R
\name{transform_restrict_age}
\alias{transform_restrict_age}
\title{Filter a variable by age}
\usage{
transform_restrict_age(lissy_files, variable, from, to)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character string with the name of the variable that should be adjusted.}
}
\value{
A list of tibbles with the adjusted variable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
'transform_restrict_age()' recodes the indicated variable to NA when age is
outside the boundaries. Cases where age is NA are also recoded to NA.
To remove rows outside of age boundaries use 'transform_filter_age()' instead.
}
