% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_subset.R
\name{transform_restrict_to_household_heads}
\alias{transform_restrict_to_household_heads}
\title{Filter a variable for household heads}
\usage{
transform_restrict_to_household_heads(lissy_files, variable)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS p-level files.}

\item{variable}{A character string with the name of the variable that should
be recoded for non-household heads.}
}
\value{
A list of tibbles with the adjusted variable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Recodes the specified variable to NA for non-household heads.
}
