% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_subset.R
\name{transform_subset_age}
\alias{transform_subset_age}
\title{Subset a sample by age}
\usage{
transform_subset_age(lissy_files, from, to)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS p-level files.}

\item{from}{An integer with the lower boundary value (included in the sample).}

\item{to}{An integer with the higher boundary value (included in the sample).}
}
\value{
A list of tibbles with the subset of rows.
}
\description{
'transform_subset_age()' subsets the rows in a LIS/LWS file so all cases have
ages between two values (both included), returning a list with the files
containing only rows where the age is between those values. I.e. a subset
of rows.
It can also be called with 'transform_filter_age()'.
}
\details{
\lifecycle{experimental}
}
