% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_weight_by_hh_size}
\alias{transform_weight_by_hh_size}
\title{Multiply by household size}
\usage{
transform_weight_by_hh_size(lissy_files)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}
}
\value{
A list of tibbles with the modified 'hwgt' variable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Multiplies the household weights ('hwgt') by the number of individuals in the
household ('nhhmem'). It is used for computations at household level.
}
