% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_zeros_to_na}
\alias{transform_zeros_to_na}
\title{Recodes zeros into missing values}
\usage{
transform_zeros_to_na(lissy_files, variable)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character string with the name of the variable that should be adjusted}
}
\value{
A list of tibbles with the adjusted variable.
}
\description{
\lifecycle{experimental}
Recodes all zeros in the selected variable into missing valuse (NAs).
}
