% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{united_states_16_lis_h}
\alias{united_states_16_lis_h}
\title{Sample Household-Level LIS Dataset (1,000 Rows)}
\format{
Data frame with 1,000 rows and 112 columns.
}
\usage{
united_states_16_lis_h
}
\description{
This dataset provides household-level data from the LIS database with 1,000 rows,
designed for exploration and user interaction. It is automatically loaded with
\code{lissyuse_locally()} whenever United States is selected along with at least one household-level
variable (e.g., "region_c").
}
\details{
The primary purpose of this dataset, and the \code{lissyuse_locally()} function, is to allow
users to build and test code locally, simulating its execution before running it on LISSY
(the LIS user interface) using the \code{lissyuse()} function.
}
\examples{
\dontrun{
united_states_16_lis_h
}
}
\keyword{datasets}
