% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{united_states_16_lis_p}
\alias{united_states_16_lis_p}
\title{Sample Individual-Level LIS Dataset (1,000 Rows)}
\format{
Data frame with 1,000 rows and 97 columns.
}
\usage{
united_states_16_lis_p
}
\description{
This dataset provides individual-level data from the LIS database with 1,000 rows,
designed for exploration and user interaction. It is automatically loaded with
\code{lissyuse_locally()} whenever United States is selected along with at least one individual-level
variable (e.g., "age").
}
\details{
The primary purpose of this dataset, and the \code{lissyuse_locally()} function, is to allow
users to build and test code locally, simulating its execution before running it on LISSY
(the LIS user interface) using the \code{lissyuse()} function.
}
\examples{
\dontrun{
united_states_16_lis_p
}
}
\keyword{datasets}
