% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us14_h_lis}
\alias{us14_h_lis}
\title{Sample Household-Level LIS Dataset - US 2014}
\format{
Data frame with 1,000 rows and 112 columns.
}
\usage{
us14_h_lis
}
\description{
Sample Household-Level LIS Dataset - US 2014
}
\examples{
\dontrun{
us14_h_lis
}
}
\keyword{datasets}
