% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us16_p_lis}
\alias{us16_p_lis}
\title{Sample Individual-Level LIS Dataset - US 2016}
\format{
Data frame with 1,000 rows and 97 columns.
}
\usage{
us16_p_lis
}
\description{
Sample Individual-Level LIS Dataset - US 2016
}
\examples{
\dontrun{
us16_p_lis
}
}
\keyword{datasets}
