% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us16_p_lws}
\alias{us16_p_lws}
\title{Sample Individual-Level LWS Dataset - US 2016}
\format{
Data frame with 2,396 rows and 133 columns.
}
\usage{
us16_p_lws
}
\description{
Sample Individual-Level LWS Dataset - US 2016
}
\examples{
\dontrun{
us16_p_lws
}
}
\keyword{datasets}
