% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us19_h_lws}
\alias{us19_h_lws}
\title{Sample Household-Level LWS Dataset - US 2019}
\format{
Data frame with 1,000 rows and 214 columns.
}
\usage{
us19_h_lws
}
\description{
Sample Household-Level LWS Dataset - US 2019
}
\examples{
\dontrun{
us19_h_lws
}
}
\keyword{datasets}
