% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us19_p_lws}
\alias{us19_p_lws}
\title{Sample Individual-Level LWS Dataset (1,000 Rows)}
\format{
Data frame with 2,273 rows and 133 columns.
}
\usage{
us19_p_lws
}
\description{
Sample Individual-Level LWS Dataset (1,000 Rows)
}
\examples{
\dontrun{
us19_p_lws
}
}
\keyword{datasets}
