% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lissyuse.r
\name{variable_selection_for_lissyuse}
\alias{variable_selection_for_lissyuse}
\title{Load the datasets for lissyuse, and automatically merge h and p level files based on the variables.}
\usage{
variable_selection_for_lissyuse(
  data_to_load,
  path_to_files,
  vars = NULL,
  database = "lis"
)
}
\arguments{
\item{data_to_load}{Character vector of dataset names to load.}

\item{path_to_files}{File path where datasets are stored.}

\item{vars}{Optional character vector of variable names to select. If \code{NULL}, all variables are considered.}

\item{database}{A character value. One of "lis" (default), "lws", or "lcs".}
}
\value{
A list containing the data frames selected.
}
\description{
Internal function to load datasets from specified files.
Depending on the variable criteria, it loads household-level, person-level,
or both datasets. If both are loaded, it merges them; otherwise, it returns
the single dataset. The merge stage may be skipped if only one dataset type is required.
}
\keyword{internal}
