% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all.values.equal}
\alias{all.values.equal}
\title{All values are equal
\lifecycle{experimental}}
\usage{
\method{all}{values.equal}(target, current)
}
\description{
Like base::all.equal(), but ignoring attributes by default.
}
\keyword{internal}
