% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_equivalisation.R
\name{apply_sqrt_equivalisation}
\alias{apply_sqrt_equivalisation}
\title{Apply Square-Root Equivalisation to a Variable}
\usage{
apply_sqrt_equivalisation(data_list, var_name, eq_scale = 0.5)
}
\arguments{
\item{data_list}{A named list of data frames, from LIS or LWS microdata.}

\item{var_name}{A string. The name of the variable to be equivalised.}

\item{eq_scale}{A numeric value between 0 and 1. The equivalence scale to apply.
Defaults to \code{0.5}, corresponding to square-root equivalisation.}
}
\value{
A list of data frames with the equivalised variable.
}
\description{
This function adjusts a household-level variable (e.g., "dhi", "hicapital")
for household size using a square-root scale equivalisation method.
It divides the variable by the number of household members raised to a power (\code{eq_scale}),
which defaults to 0.5 (i.e., the square root equivalence scale).
}
\details{
The function assumes that the dataset contains a column named \code{nhhmem} representing
the number of household members. If this column is missing, the function stops with an error.

It also issues a warning if the target variable is detected as an individual-level variable,
since equivalisation is typically applied to household-level measures.
}
\examples{
\dontrun{
# Import data, ideally at the level of the variable of interest.
data_hhd <- lissyrtools::lissyuse("it", vars = c("dhi"), from = 2016)

# Default case: square-root equivalisation
data_hhd[1]  \%>\%
 purrr::map(~ .x \%>\% filter(!is.na(dhi))) \%>\%
 apply_sqrt_equivalisation("dhi")
}
}
