% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ccyy_to_cname}
\alias{ccyy_to_cname}
\title{Get the country name from the `ccyy``}
\usage{
ccyy_to_cname(ccyy)
}
\arguments{
\item{ccyy}{A character vector of ccyy's}
}
\value{
A character vector with the corresponding country names.
}
\description{
Get the country name from the `ccyy``
}
\keyword{internal}
