% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ccyy_to_yyyy}
\alias{ccyy_to_yyyy}
\title{Get the year from the `ccyy``}
\usage{
ccyy_to_yyyy(ccyy)
}
\arguments{
\item{ccyy}{A character vector of ccyy's}
}
\value{
A numeric vector with the corresponding years.
}
\description{
Get the year from the `ccyy``
}
\keyword{internal}
