% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_for_lissyuse.R
\name{check_empty_data}
\alias{check_empty_data}
\title{Checks for Empty Data Input}
\usage{
check_empty_data(data, lws = FALSE)
}
\arguments{
\item{data}{A character vector.}

\item{lws}{Logical. If \code{TRUE}, suggests loading all LWS countries; otherwise, suggests LIS countries.}
}
\value{
Stops with an informative error message if \code{data} is \code{NULL}.
}
\description{
Internal helper to ensure the \code{data} argument is not \code{NULL} when required. Suggests loading all countries if input is missing.
}
\keyword{internal}
