% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_for_lissyuse.R
\name{check_iso2}
\alias{check_iso2}
\title{Checks for Invalid iso2 codes.}
\usage{
check_iso2(data, lws = FALSE)
}
\arguments{
\item{data}{A character vector.}

\item{lws}{Logical.}
}
\value{
Stops if all iso2 codes in \code{data} are invalid, and it warns if only some are invalid.
}
\description{
Internal helper to ensure the iso2 codes in argument \code{data} correspond to valid iso2 codes.
}
\keyword{internal}
