% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_count.R
\name{compute_weighted_count}
\alias{compute_weighted_count}
\title{Compute (Weighted) Counts or Percentages for a Categorical Variable}
\usage{
compute_weighted_count(var, wgt = NULL, na.rm = FALSE, percent = FALSE)
}
\arguments{
\item{var}{A column refering to one of the categorical variables in a LIS or LWS data frame.}

\item{wgt}{A numeric vector of weights (e.g., .x$hpopwgt, .x$pwgt). Must be the same length as \code{x}.}

\item{na.rm}{Logical; if \code{TRUE}, missing values in \code{x} and \code{w} are removed before computation. Default is \code{FALSE}.}

\item{percent}{Logical; if \code{TRUE}, computes weighted (or non-weighted) percentages.}
}
\value{
A numeric vector, with category labels as names.
}
\description{
Compute (Weighted) Counts or Percentages for a Categorical Variable
}
\examples{
\dontrun{
data <- lissyrtools::lissyuse(data = "de20", vars = c("dhi", "age", "educ"))
compute_weighted_count(data$de20$educ, na.rm = TRUE)
compute_weighted_count(data$de20$educ, percent = TRUE)
compute_weighted_count(data$de20$educ, na.rm = TRUE, percent = TRUE)
compute_weighted_count(data$de20$educ, data$de20$ppopwgt, na.rm = TRUE, percent = TRUE)
compute_weighted_count(data$de20$educ, data$de20$ppopwgt, percent = TRUE)
}
}
\keyword{internal}
