% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean.R
\name{compute_weighted_mean}
\alias{compute_weighted_mean}
\title{Compute (Weighted) Means for a Variable}
\usage{
compute_weighted_mean(var, wgt = NULL, na.rm = TRUE)
}
\arguments{
\item{var}{A column refering to one of the variables in a LIS or LWS data frame.}

\item{wgt}{A numeric vector of weights (e.g., .x$hpopwgt, .x$pwgt). Must be the same length as \code{x}.}

\item{na.rm}{Logical; if \code{TRUE}, missing values in \code{x} and \code{w} are removed before computation. Default is \code{TRUE}.}
}
\value{
A numeric vector.
}
\description{
Compute (Weighted) Means for a Variable
}
\examples{
\dontrun{
data <- lissyrtools::lissyuse(data = "de20", vars = c("dhi", "age", "educ"))
compute_weighted_mean(data$de20$age, na.rm = TRUE)
compute_weighted_mean(data$de20$dhi, data$de20$hwgt)
}
}
