% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_with_warnings}
\alias{data_with_warnings}
\title{Variable notes}
\format{
A data frame with 6 variables:
}
\usage{
data_with_warnings
}
\description{
A data frame indicating whether a given variable in one of the LIS/LWS datasets has a note or not for a specific year.
}
\keyword{datasets}
