% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{LIS and LWS datasets}
\format{
A data frame with 7 variables:
}
\usage{
datasets
}
\description{
A data frame containing all the LIS and LWS datasets
that can be used through LISSY, including some of their characteristics.
}
\keyword{datasets}
