% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_for_lissyuse.R
\name{invalid_ccyy_pairs}
\alias{invalid_ccyy_pairs}
\title{Checks for Invalid ccyy.}
\usage{
invalid_ccyy_pairs(data, lws = FALSE)
}
\arguments{
\item{data}{A character vector.}

\item{lws}{Logical.}
}
\value{
Stops if all ccyy codes in \code{data} are invalid, and it warns if only some are invalid.
}
\description{
Internal helper to ensure the ccyy pairs in argument \code{data} are valid.
}
\keyword{internal}
