% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lissy_map.R
\name{lissy_map}
\alias{lissy_map}
\title{Wrap over 'purrr::imap()' for lissy files}
\usage{
lissy_map(lissy_files, .f)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS p-level files.}

\item{.f}{A function, formula, or vector (not necessarily atomic) as in
the argument from purrr::map with the same name.}
}
\value{
A list like lissy_files
}
\description{
A wrap over 'purrr::imap()' that keeps the lissy attributes on the output
list.
}
