% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lissyrtools_package_documentation.R
\docType{"_PACKAGE"}
\name{lissyrtools}
\alias{lissyrtools}
\title{lissyrtools: Functions that can be used within LISSY}
\description{
The lissyrtools provides three types of functions:
read, transform, and print. These can be chained using
'magrittr' pipes ('\%>\%') in 'tidyverse' style.
}
\section{read functions}{

Read and merge LISSY files.
}

\section{transform functions}{

Perform data processing/cleaning processes.
}

\section{print functions}{

Compute estimates.
}

