% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lissyuse.r
\name{lissyuse}
\alias{lissyuse}
\title{Load data easily and efficiently with lissyuse}
\usage{
lissyuse(
  data = NULL,
  vars = NULL,
  subset = NULL,
  from = NULL,
  to = NULL,
  lws = FALSE
)
}
\arguments{
\item{data}{}

\item{vars}{}

\item{subset}{}

\item{from}{}

\item{to}{}

\item{lws}{}
}
\value{
A list
}
\description{
\code{lissyuse()} enables LISSY R-users to quickly import whole series of data from a specific country and/or within any time period.
}
\examples{
# LIS # 
}
