% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_all_lissy_files}
\alias{print_all_lissy_files}
\title{Print all available files}
\usage{
print_all_lissy_files(database)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Display all available LISSY files.
}
\examples{
\dontrun{
print_all_lissy_files()
}
}
\keyword{internal}
