% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_gini}
\alias{print_gini}
\title{Print Gini}
\usage{
print_gini(
  lissy_files,
  variable,
  weight = NULL,
  na.rm = FALSE,
  files_level = NULL,
  variable_level = NULL
)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files.}

\item{variable}{A character string indicating the aggregate for which the indicator needs to be computed.}

\item{weight}{A string with the name of the variable in 'file' that should be used as sample weights.}

\item{na.rm}{A boolean. Indicates if NAs should be ignored. Defaults to FALSE.}

\item{files_level}{A string indicating the level of the file. Valid inputs are:
'household', 'h', 'person' or 'p'. If NULL (default), the file level will
be retrived from the 'lissy_files' attributes.}

\item{variable_level}{Level of the variable. Should be either 'household', 'h', 'person' or 'p'.
If NULL (default), the function will try to guess the level of the variable.
This is done by comparing the value in 'variable' with pre-set lists of variables.}
}
\value{
A named numeric vector with the Gini coefficients.
}
\description{
\lifecycle{experimental}
Computes and displays the Gini coefficient.
}
\examples{
\dontrun{
lissy_datasets <- read_lissy_files(c("fr84h", "fr94h", "fr10h"))
print_gini(lissy_datasets = lissy_datasets, variable = "dhi", na.rm = TRUE)
}
}
