% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_equivalise_oecd}
\alias{transform_equivalise_oecd}
\title{Equivalise with the OECD scale}
\usage{
transform_equivalise_oecd(
  lissy_files,
  variable,
  value_other_adults = 0.7,
  value_children = 0.5
)
}
\arguments{
\item{lissy_files}{A list of LIS or LWS files}

\item{variable}{A character vector of length one with the indicator that needs to be transformed.}

\item{value_other_adults}{A real number. Defaults to 0.7. The value assigned to the
second to last adults in the household.}

\item{value_children}{A real number. Defaults to 0.7. The value assigned to children
in the household.}
}
\value{
A list of tibbles with the transformed variable.
}
\description{
\lifecycle{experimental}
Scales a variable using a weight for the adults (excluding the first one) and
children.
}
\details{
Throws a warning if the function is applied to a person-level variable.
}
\examples{
\dontrun{
lissy_files <- read_lissy_files(c("fr84h", "fr94h", "fr10h"))
transform_equivalise_oecd(list_files = lissy_files, variable = "dhi")
}
}
