% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels_notes_surveys.R
\name{variable_has_note}
\alias{variable_has_note}
\title{Verify if a given variable has a note, for a selected group of countries, in LIS or LWS databases.}
\usage{
variable_has_note(variable, iso2, database = "lis", ...)
}
\arguments{
\item{variable}{A unit-length character vector containing specified LIS/LWS variables.}

\item{iso2}{A character vector with valid iso2 codes of countries present in LIS/LWS.}

\item{database}{A character value. One of "lis" (default), "lws", or "lcs".}
}
\value{
A list, made of character vectors. Each elements corresponds to a country in LIS or LWS databases.
}
\description{
Verify if a given variable has a note, for a selected group of countries, in LIS or LWS databases.
}
\examples{
\dontrun{
variable_has_note(variable = "area_c", iso2 = "br")
variable_has_note(variable = "basb", iso2 = c("fr", "de", "us", "uk"), database = "lws")

To check the notes on METIS go to: https://www.lisdatacenter.org/frontend#/home, and select a database and a given country:.
Afterwards, head to 'RESULTS' > 'Dataset information' > 'Code Books'.
}
}
